+++ app/src/main/java/mb28/crysongs/ui/more_pages/FoldersPage.kt
@@ -1,0 +1,100 @@
+package mb28.crysongs.ui.more_pages
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.material3.FilledTonalIconButton
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.runtime.toMutableStateList
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import mb28.crysongs.folders
+import mb28.crysongs.icons.arrow_back
+import mb28.crysongs.icons.download
+import mb28.crysongs.icons.folder
+import mb28.crysongs.playerQuery
+import mb28.crysongs.tracks
+import mb28.crysongs.ui.other.EasySegmentedListItem
+import mb28.crysongs.ui.other.TrackTile
+import mb28.crysongs.updateDisplayQuery
+
+@Composable
+fun FoldersPage() {
+    var folderView by remember { mutableStateOf(false) }
+    var clickedFolderPath by remember { mutableStateOf("") }
+    LazyColumn(
+        contentPadding = PaddingValues(top = 130.dp, bottom = 200.dp),
+    ) {
+        item {
+            Text(
+                if (folderView) clickedFolderPath.substring(
+                    clickedFolderPath.lastIndexOf('/') + 1)
+                    else "Folders (${folders.count()})",
+                fontSize = 36.sp,
+                textAlign = TextAlign.Center,
+                modifier = Modifier.fillMaxWidth()
+            )
+            Spacer(Modifier.height(33.dp))
+        }
+
+        item {
+            Row(
+                Modifier.fillMaxWidth().padding(top = 40.dp, bottom = 10.dp, start = 10.dp, end = 10.dp),
+                horizontalArrangement = Arrangement.Start
+            ) {
+                if (folderView) {
+                    FilledTonalIconButton(
+                        { folderView = false }
+                    ) {
+                        Icon(arrow_back, null)
+                    }
+                }
+            }
+        }
+
+        if (folderView) {
+            val folderTracks = tracks.takeWhile {
+                it.path.startsWith(clickedFolderPath)
+            }
+            val count = folderTracks.count()
+            items(count) { i ->
+                TrackTile(
+                    folderTracks[i],
+                    i, count,
+                    resetQueryOnClick = false
+                ) {
+                    playerQuery = folderTracks.toMutableStateList()
+                    updateDisplayQuery()
+                }
+            }
+        }
+        else {
+            val count = folders.count()
+            items(count) { i ->
+                val f = folders[i]
+                EasySegmentedListItem(
+                    folder,
+                    f.substring(f.lastIndexOf('/') + 1),
+                    i, count,
+                    Modifier.padding(horizontal = 10.dp)
+                ) {
+                    clickedFolderPath = f
+                    folderView = true
+                }
+            }
+        }
+    }
+}